feat: Trident v2.1 - 7-9月稳定期锚定框架
- 新增7-9月稳定期作为基准期的预测逻辑
- 只预测最后一年10月-次年6月的月度均值
- 维修后作为新基础设施，重新计算趋势
- 近5年维修检测，避免早期假维修干扰
- v2.1在50样本对比中击败TimeMixer (68%胜率)

新增文件:
- trident_v21.py: v2.1主实现
- trident_v2.py: v2实现
- test_v21_vs_timemixer.py: 对比测试脚本
- results/v21_vs_timemixer/: 测试结果

+++ 03-实验与实现/src/baseline_491_experiment.py
@@ -237,8 +237,9 @@ def lstm_baseline(train_df, test_df):
         
         tf.get_logger().setLevel('ERROR')
         
-        train_values = train_df['tqi_value'].values
-        test_values = test_df['tqi_value'].values
+        # 修复: 强制转换为float64，避免object dtype错误
+        train_values = train_df['tqi_value'].values.astype(np.float64)
+        test_values = test_df['tqi_value'].values.astype(np.float64)
         
         if len(train_values) < 50:
             return {'mae': float('nan'), 'rmse': float('nan'), 'mape': float('nan')}
@@ -256,7 +257,8 @@ def lstm_baseline(train_df, test_df):
             for i in range(len(data) - seq_len):
                 X.append(data[i:i+seq_len])
                 y.append(data[i+seq_len])
-            return np.array(X), np.array(y)
+            # 修复: 强制转换为float64，避免object dtype错误
+            return np.array(X, dtype=np.float64), np.array(y, dtype=np.float64)
         
         X_train, y_train = create_sequences(train_normalized, seq_length)
         if len(X_train) < 10:
